requirements.txt
audioread==3.0.1
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.2
chatterbox-tts==0.1.2
click==8.2.1
conformer==0.3.2
decorator==5.2.1
diffusers==0.29.0
distlib==0.3.9
einops==0.8.1
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
hf-xet==1.1.5
huggingface-hub==0.33.2
identify==2.6.12
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
msgpack==1.1.1
networkx==3.5
nodeenv==1.9.1
numba==0.61.2
numpy==2.2.6
onnx==1.18.0
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
pre_commit==4.2.0
protobuf==6.31.1
pycparser==2.22
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
resemble-perth==1.0.1
s3tokenizer==0.1.7
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
setuptools==80.9.0
soundfile==0.13.1
soxr==0.5.0.post1
sympy==1.13.1
threadpoolctl==3.6.0
tokenizers==0.20.3
torch==2.6.0
torchaudio==2.6.0
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.14.0
urllib3==2.5.0
virtualenv==20.31.2
Werkzeug==3.1.3
zipp==3.23.0
